Fix the onCreate stuff
We check if the LockScreen has been fired by the HOME button,
so if is, we just redirect to the launcher.

We need to change this code to the one from onNewIntent to check if there's already a launcher
selected, etc etc.

The only stuff not working with this,is that the real launcher home key bindings stop working....

And now we should find a way to handle the SCREEN_OFF and/or SCREEN_ON
to load the lockscreen.....